Use pkg-config to find SDL on Linux
This gets rid of the "-L" flag which results in our local deps not being
picked up if these deps exist as system libs. Will need to solve that in
a better way later on. For now, it should result in our static libs being
actually used in the buildbot.