Add support for changing current core option values programmatically
This is a hack. The libretro API does not actually support this. We
achieve this by replacing all current values with a bogus one to force
the frontend to forget the currently selected value, since it doesn't
match the bogus one. We then submit the correct values again, but with
the default value set to the value we want to switch to. This forces
the frontend to switch to that value because the bogus value is now
gone and thus not a valid value anymore. Finally, we submit the values
again but with the initial default value (we only want to change the
current value, not set a new default.) The frontend will not switch the
current value, as the values themselves have not changed, just the
default has.

RetroArch seems to be well-behaved here and does the correct thing.
Other frontends might not play ball though.