Add BUNDLED_GLIB make var for bundled glib. enable it on Unix by default
Using a statically linked glib fixes a crash/hang on Linux when the core
is unloaded and then loaded again in RetroArch. There seems to be a
conflict with how fluidsynth uses glib and how RetroArch itself uses it.
Not sure what's going on exactly, but using our own bundled glib fixes
it.