Build cmake deps with CMAKE_CXX_STANDARD=17
Only important for fmt really, since it's best if it's built using the
same C++ standard as the code that uses it. But use it for all other
cmake-based deps as well for consistency. Who knows if there's any
hidden ABI issues when some code is built targetting a different
standard.