Add d-pad mouse emulation, and some nits
Now that there's no mouse emulation toggle in the core options, we need
to do mouse emulation using the d-pad when using a 4-axis joystick,
since the right analog stick is needed by the joystick.

Previously, we couldn't have both; we had to choose between mouse
emulation and not having a right stick, or having a right stick but no
mouse emulation. Now we can have both since the d-pad is free on analog
Joysticks.