Remove all git submodules and replace them with subrepos
I've had enough with submodules. The camel that broke the water's drop,
or however the fuck the saying goes, is that a git clone --recursive
ends up checking out the wrong submodule commit for no apparent reason,
even though the submodule is configured to track a specific commit.

Hopefully subrepos will behave in a way that's not brain damaged.