Add core option for disabling frame duping
When using the Vulkan driver in RetroArch, frame duping can result in a
black screen until the next input that causes a new frame to be
rendered. This happens when RetroArch does a video driver reinit. For
example when switching RetroArch from fullscreen to windowed or vice
versa. Or when the core requests a reinit through
RETRO_ENVIRONMENT_SET_SYSTEM_AV_INFO.