Core option to improve mouse speed for games with multiple video modes
This is just a hack and kinda dumb, but it works in many games. We just
use a factor of 0.85 for the vertical sensitivity if the current video
mode is based on 400 line output (like 320x200, 640x400, etc.) and a
factor of 2 if it's based on 480 line output.

Disabled by default.