Unify core options with .conf file INI settings and runtime conf changes
This allows core options to reflect configuration settings from the
loaded .conf file as well as configuration changes done at runtime
through dosbox commands such as "config -set" or "cycles".

There's two settings to choose from, "lock" and "disable". The former
will lock the corresponding core option to a single value and label it
as "Locked". The latter will not change the values of the option but
will hide it in the core options UI (if supported by the frontend) and
ignore all future changes to the core option.

Unfortunately, this requires changing the ID of a lot of the existing
core options to match their .conf property counterparts. This will
invalidate .opt files of current users. The changed options are:

  machine_type -> machine
  voodoo_memory_size -> voodoomem
  memory_size -> memsize
  cpu_core -> core
  cpu_type -> cputype
  aspect_correction -> aspect
  joystick_timed -> timed
  sblaster_type -> sbtype
  sblaster_base -> sbbase
  sblaster_irq -> irq
  sblaster_dma -> dma
  sblaster_hdma -> hdma
  sblaster_opl_mode -> oplmode
  sblaster_opl_emu -> oplemu
  mpu_type -> mpu401
  midi_driver -> mididevice
  midi_port -> midiconfig