Improve audio frame pacing and latency
The core is only uploading a truncated amount of audio frames to the
frontend. This has two drawbacks. It results in an audio and video frame
stutter every so often (could be minutes apart). It also introduces
audio lag because samples are piling up in the dosbox mixer.

Fix this by uploading all available audio to the frontend, flushing the
dosbox mixer every frame. This should also make the average sample count
per frame match what the frontend expects.