Canonicalize game path on retro_load_game()
This gets rid of any "." and ".." parts of the path, resolves symlinks
and makes sure the path is absolute.

This ensures that overlay mounting can't fail if the path doesn't
contain any directory separators, which happens when launching the core
from the command line and the content is a file in the current
directory. Since the overlay path is always absolute, dosbox would fail
to mount the overlay because the base directory path of the C drive is
relative.