Replace fluidsynth with fluidsynth-sans-glib
fluidsynth-sans-glib is a fork of fluidsynth that gets rid of the glib
dependency. glib is very problematic on Windows, since it can crash
frontends on dlclose and dlopen. Windows doesn't always unload libraries
on dlclose, which causes glib to crap its pants due to persistent
statics that aren't being cleaned up.

Getting rid of glib also allows us to get rid of libffi and meson.

The downside is that we can't build fluidsynth with libinstpatch support
anymore, since it also depends on glib. So we lose support for DLS and
and GigaSampler soundfonts. Fortunately, everybody uses SF2 soundfonts,
so it's not a huge problem.