Fix library linking order and missing libsndfile
Not sure why this is working fine on some architectures while on others
we get "undefined symbol" when trying to load the core. In any event,
the linking order was wrong and libsndfile not getting linked should
have resulted in errors on all systems. Who knows what was going on.