Add option for clamping minimal mouse speed
Forces very slow mouse movements to be registered as slightly faster.
This can be useful in games that don't register mouse movements that are
slower than a certain threshold. Flight of the Amazon Queen is an
example of such a game. Do not enable this option in games that don't
have this problem. It makes mouse movement worse.

What this does is simply clamp the values passed to Mouse_CursorMoved()
so that they can't be between -1 and 1. It doesn't fix the buggy games,
but can make things a bit less annoying.